+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 v0.6.9 (TBD)
 --------------------
  - [Correlation] Catch crash if there is insufficient data for a correlation calculation
- - [Machine Learning] Allow user to load saved models
- - [Query] Fix [issue 20](https://github.com/cutright/DVH-Analytics/issues/20)
+ - [Modeling] Allow user to load saved models (Multi-Variable and Machine Learning)
+ - [Query] Fix for [issue 20](https://github.com/cutright/DVH-Analytics/issues/20)
  - [Query] Corrected physician count of query